requirements.txt
einops==0.6.0
numpy==1.23.5
pandas==1.3.1
path==16.6.0
path.py==12.5.0
pathtools==0.1.2
Pillow==9.4.0
PyYAML 
timm==0.3.2
tqdm==4.64.1

